feat: Busca por demandas criadas e resolvidas

+++ adi-gojira/pkg/gjmodels/pagination.go
@@ -4,5 +4,5 @@ type Pagination struct {
 	Expand     string `json:"expand"`
 	StartAt    uint   `json:"startAt"`
 	MaxResults uint   `json:"maxResults"`
-	Total      uint64 `json:"total"`
+	Total      uint   `json:"total"`
 }
